PRSNT: Modifica parametri di ICompassView.setHeading().
Il tipo del valore in gradi dell'orientamento dell'utente passato a setHeading()
viene modificato da float a double per allinearsi al resto dell'applicazione.

Related to SHANDROID-63

+++ serleena/app/src/main/java/com/kyloth/serleena/presentation/ICompassView.java
@@ -58,7 +58,7 @@ public interface ICompassView {
 	 *                modulo 360.0.
 	 * @since 1.0
 	 */
-	void setHeading(float heading);
+	void setHeading(double heading);
 
 	/**
 	 * Lega un presenter alla vista.
